backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
python-multipart>=0.0.6
pydantic>=2.4.0
pydantic-settings>=2.0.0
typing-extensions>=4.12.0
